+++ examples/spinner.ts
@@ -1,10 +1,10 @@
-import { spinner } from "../src/utils/prompt";
+import { spinner } from "../src/utils";
 
 async function main() {
-  const s = spinner();
-  s.start("Creating project...");
+  const spin = spinner();
+  spin.start("Creating project...");
   await new Promise((resolve) => setTimeout(resolve, 1000));
-  s.stop("Project created!");
+  spin.stop("Project created!");
 }
 
 main();
